# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce API
  version: "1.0.0"
servers:
  - url: http://localhost:8080

paths:
  /users/register:
    post:
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created

  /users/login:
    post:
      summary: Login (simple)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login OK

  /orders:
    post:
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created


    get:
      summary: List orders
      responses:
        "200":
          description: Orders list

  /orders/{id}:
    get:
      summary: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail

components:
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }

    CreateOrderRequest:
      type: object
      required: [userId, items]
      properties:
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    OrderItem:
      type: object
      required: [sku, name, quantity, price]
      properties:
        sku: { type: string }
        name: { type: string }
        quantity: { type: integer }
        price: { type: number }